refactor(rfc64): centralize runtime evidence and failures
