+++ devnet/rfc64-cp1-private-swm-recovery/launch-live.ts
@@ -5,7 +5,7 @@ import {
   buildGate2RuntimeManifestV1,
   installGate2RuntimeLaunchReceiptV1,
   runGate2CleanRuntimeBuildV1,
-} from '../rfc64-gate2-multi-asset-completeness/runtime-provenance.ts';
+} from '../rfc64-runtime-provenance.mts';
 
 const repoRoot = resolve(import.meta.dirname, '../..');
 const sourceCommit = readCleanRepositoryHead(repoRoot);
