+++ devnet/rfc64-cp1-private-swm-recovery/run.ts
@@ -36,7 +36,7 @@ import {
 } from '../rfc64-gate2-multi-asset-completeness/agent-child.js';
 import {
   consumeGate2RuntimeLaunchReceiptV1,
-} from '../rfc64-gate2-multi-asset-completeness/runtime-provenance.ts';
+} from '../rfc64-runtime-provenance.mts';
 import {
   assertGate2HarnessReadyV1,
   assertGate2HarnessSourceStateV1,
