+++ devnet/rfc64-cp1-private-swm-recovery/tsconfig.json
@@ -14,7 +14,7 @@
     "*.ts",
     "../rfc64-gate2-multi-asset-completeness/agent-child.ts",
     "../rfc64-gate2-multi-asset-completeness/model.ts",
-    "../rfc64-gate2-multi-asset-completeness/runtime-provenance.ts",
+    "../rfc64-runtime-provenance.mts",
     "../rfc64-gate2-multi-asset-completeness/two-agent-harness.ts",
     "../rfc64-persistence-lifecycle/evidence.ts",
     "../rfc64-persistence-lifecycle/process-lifecycle.ts"
