+++ devnet/rfc64-cp3-private-provider-failover/tsconfig.json
@@ -15,7 +15,7 @@
     "../rfc64-gate2-multi-asset-completeness/agent-child.ts",
     "../rfc64-gate2-multi-asset-completeness/finalized-vm-harness-runtime.ts",
     "../rfc64-gate2-multi-asset-completeness/model.ts",
-    "../rfc64-gate2-multi-asset-completeness/runtime-provenance.ts",
+    "../rfc64-runtime-provenance.mts",
     "../rfc64-gate2-multi-asset-completeness/two-agent-harness.ts",
     "../rfc64-persistence-lifecycle/evidence.ts",
     "../rfc64-persistence-lifecycle/process-lifecycle.ts"
