+++ packages/agent/devnet/rfc64-private-catalog/clean-launch.ts
@@ -2,35 +2,35 @@ import { execFileSync } from 'node:child_process';
 
 import { readCleanRepositoryHead } from '../../../../devnet/rfc64-persistence-lifecycle/evidence.ts';
 import {
-  GATE2_RUNTIME_PACKAGE_CLOSURE,
-  assertGate2RuntimeManifestEqualV1,
-  buildGate2RuntimeManifestV1,
-  runGate2CleanRuntimeBuildV1,
-  type Gate2RuntimeManifestV1,
-} from '../../../../devnet/rfc64-gate2-multi-asset-completeness/runtime-provenance.ts';
+  RFC64_RUNTIME_PACKAGE_CLOSURE_V1,
+  assertRuntimeManifestEqualV1,
+  buildRuntimeManifestV1,
+  runCleanRuntimeBuildV1,
+  type RuntimeManifestV1,
+} from '../../../../devnet/rfc64-runtime-provenance.mts';
 import { runRfc64PrivateGateArtifactLifecycleV1 } from './gate-artifact.mjs';
 
 interface Rfc64PrivateGateCleanLaunchDependenciesV1 {
   readonly buildRuntimeManifest:
-    (repoRoot: string, sourceRevision: string) => Readonly<Gate2RuntimeManifestV1>;
+    (repoRoot: string, sourceRevision: string) => Readonly<RuntimeManifestV1>;
   readonly readCleanSourceRevision: (repoRoot: string) => string;
   readonly resolveSourceRevision: (repoRoot: string) => string;
   readonly runCleanRuntimeBuild: (repoRoot: string) => void;
 }
 
 const DEFAULT_DEPENDENCIES: Rfc64PrivateGateCleanLaunchDependenciesV1 = Object.freeze({
-  buildRuntimeManifest: buildGate2RuntimeManifestV1,
+  buildRuntimeManifest: buildRuntimeManifestV1,
   readCleanSourceRevision: readCleanRuntimeBuildSourceRevisionV1,
   resolveSourceRevision: (repoRoot) => execFileSync('git', ['rev-parse', 'HEAD'], {
     cwd: repoRoot,
     encoding: 'utf8',
     stdio: ['ignore', 'pipe', 'ignore'],
   }).trim(),
-  runCleanRuntimeBuild: runGate2CleanRuntimeBuildV1,
+  runCleanRuntimeBuild: runCleanRuntimeBuildV1,
 });
 
 const RUNTIME_BUILD_INPUT_PATHS = Object.freeze([
-  ...GATE2_RUNTIME_PACKAGE_CLOSURE.map(({ path }) => (
+  ...RFC64_RUNTIME_PACKAGE_CLOSURE_V1.map(({ path }) => (
     path.endsWith('/dist') ? path.slice(0, -'/dist'.length) : path
   )),
   'package.json',
@@ -47,7 +47,7 @@ export async function runRfc64PrivateGateFromCleanBuildV1(input: {
   readonly artifactPath: string;
   readonly dependencies?: Partial<Rfc64PrivateGateCleanLaunchDependenciesV1>;
   readonly execute: (input: {
-    readonly runtimeManifest: Readonly<Gate2RuntimeManifestV1>;
+    readonly runtimeManifest: Readonly<RuntimeManifestV1>;
     readonly sourceRevision: string;
   }) => Promise<Record<string, unknown>>;
   readonly repoRoot: string;
@@ -81,7 +81,7 @@ export async function runRfc64PrivateGateFromCleanBuildV1(input: {
         sourceRevision,
         dependencies.readCleanSourceRevision(input.repoRoot),
       );
-      assertGate2RuntimeManifestEqualV1(
+      assertRuntimeManifestEqualV1(
         dependencies.buildRuntimeManifest(input.repoRoot, sourceRevision),
         runtimeManifest,
       );
