+++ packages/agent/src/rfc64/public-catalog-reconciliation-failure-v1.ts
@@ -94,17 +94,6 @@ export interface Rfc64PublicCatalogReconciliationFailureV1 {
   readonly causeCode?: string;
 }
 
-/** Result of the most recently started scheduler attempt for one exact head. */
-export interface Rfc64CatalogReconciliationAttemptFailureV1 {
-  readonly catalogHeadDigest: Digest32V1;
-  readonly terminalReason: Rfc64CatalogReconciliationTerminalReasonV1 | null;
-  readonly errorName: string;
-  readonly errorCode: string | null;
-}
-
-/** Monotonic process-local identity for one execution-time reconciliation attempt. */
-export type Rfc64CatalogReconciliationAttemptTokenV1 = number;
-
 /** Hard process-memory bound for distinct terminal receiver failures. */
 export const RFC64_PUBLIC_CATALOG_RECONCILIATION_FAILURE_MAX_ENTRIES_V1 = 128;
 
@@ -117,32 +106,9 @@ const STABLE_ERROR_TOKEN_V1 = /^[A-Za-z][A-Za-z0-9._:-]*$/;
  */
 export class Rfc64PublicCatalogReconciliationFailureRegistryV1 {
   readonly #failures = new Map<Digest32V1, Rfc64PublicCatalogReconciliationFailureV1>();
-  readonly #currentAttemptFailures = new Map<
-    Digest32V1,
-    Rfc64CatalogReconciliationAttemptFailureV1
-  >();
-  readonly #currentAttemptTokens = new Map<
-    Digest32V1,
-    Rfc64CatalogReconciliationAttemptTokenV1
-  >();
-  #attemptSequence = 0;
 
-  /** Start a new semantic attempt without changing immutable diagnostic history. */
-  beginAttempt(catalogHeadDigest: Digest32V1): Rfc64CatalogReconciliationAttemptTokenV1 {
-    const token = ++this.#attemptSequence;
-    evictOldestWhenFullV1(this.#currentAttemptTokens);
-    this.#currentAttemptTokens.set(catalogHeadDigest, token);
-    this.#currentAttemptFailures.delete(catalogHeadDigest);
-    return token;
-  }
-
-  /** Scheduler-terminal callback sink. The first failure for one head wins. */
-  record(
-    catalogHeadDigest: Digest32V1,
-    error: unknown,
-    terminalReason: Rfc64CatalogReconciliationTerminalReasonV1 | null = null,
-    attemptToken?: Rfc64CatalogReconciliationAttemptTokenV1,
-  ): void {
+  /** Diagnostic-history sink. The first failure for one head wins. */
+  record(catalogHeadDigest: Digest32V1, error: unknown): void {
     const errorName = stableErrorNameV1(error);
     const errorCode = stableErrorCodeV1(error);
     const causeCode = stableImmediateCauseCodeV1(error);
@@ -155,39 +121,6 @@ export class Rfc64PublicCatalogReconciliationFailureRegistryV1 {
         ...(causeCode === null ? {} : { causeCode }),
       }));
     }
-    const currentAttemptToken = this.#currentAttemptTokens.get(catalogHeadDigest);
-    if (
-      (currentAttemptToken !== undefined && currentAttemptToken !== attemptToken)
-      || (currentAttemptToken === undefined && attemptToken !== undefined)
-    ) {
-      return;
-    }
-    evictOldestWhenFullV1(this.#currentAttemptFailures);
-    this.#currentAttemptFailures.set(catalogHeadDigest, Object.freeze({
-      catalogHeadDigest,
-      terminalReason,
-      errorName,
-      errorCode,
-    }));
-  }
-
-  /** Clear only the exact current attempt; stale successes cannot erase a newer failure. */
-  completeAttempt(
-    catalogHeadDigest: Digest32V1,
-    attemptToken: Rfc64CatalogReconciliationAttemptTokenV1,
-  ): void {
-    if (this.#currentAttemptTokens.get(catalogHeadDigest) !== attemptToken) return;
-    this.#currentAttemptFailures.delete(catalogHeadDigest);
-    this.#currentAttemptTokens.delete(catalogHeadDigest);
-  }
-
-  /** Release only the exact terminal attempt token; retain its failure result. */
-  endAttempt(
-    catalogHeadDigest: Digest32V1,
-    attemptToken: Rfc64CatalogReconciliationAttemptTokenV1,
-  ): void {
-    if (this.#currentAttemptTokens.get(catalogHeadDigest) !== attemptToken) return;
-    this.#currentAttemptTokens.delete(catalogHeadDigest);
   }
 
   read(catalogHeadDigest: Digest32V1): Rfc64PublicCatalogReconciliationFailureV1 | null {
@@ -195,31 +128,13 @@ export class Rfc64PublicCatalogReconciliationFailureRegistryV1 {
     return failure === undefined ? null : failure;
   }
 
-  readCurrentAttempt(
-    catalogHeadDigest: Digest32V1,
-  ): Rfc64CatalogReconciliationAttemptFailureV1 | null {
-    return this.#currentAttemptFailures.get(catalogHeadDigest) ?? null;
-  }
-
   clear(): void {
     this.#failures.clear();
-    this.#currentAttemptFailures.clear();
-    this.#currentAttemptTokens.clear();
   }
 
   /** Internal test/diagnostic bound assertion; never exposed on DKGAgent. */
   get size(): number {
     return this.#failures.size;
-  }
-
-  /** Internal test/diagnostic bound assertion; never exposed on DKGAgent. */
-  get currentAttemptFailureSize(): number {
-    return this.#currentAttemptFailures.size;
-  }
-
-  /** Internal test/diagnostic bound assertion; never exposed on DKGAgent. */
-  get currentAttemptTokenSize(): number {
-    return this.#currentAttemptTokens.size;
   }
 }
